First Commit. Developed objects for LifeForm, and decendants, Animal, Person. Developed objects for the world, for time, for sleep and hunger. Test version spawns a single person that gets tired, falls aslep, and starves to death.
Need to add a per hour time granulation for events to better sync the
affects of time-of-day on restfulness and hunger.

Then we will move to the second tier of Maslow's heirarchy of Needs and add
a sense of security, and the need to have a home. Security will affect
sleep.